+++ source/mocassin.f90
@@ -44,9 +44,9 @@ program MoCaSSiN
 
     type(grid_type) :: grid3D(maxGrids)       ! the 3D Cartesian  grid
 
-    real            :: test                   ! test
+    real(kind=dp)            :: test                   ! test
     integer         :: i, iGrid               ! allocation error status
-    real, dimension(2) :: timing              ! cputimer 
+    real(kind=dp), dimension(2) :: timing              ! cputimer 
     integer         :: nhours, nminutes, nseconds
 
     call cpu_time(timing(1))
